api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.7.4"
Django = "2.1.5"
django-cors-headers = "2.4.0"
djangorestframework = "3.8.2"
ExifRead = "2.1.2"
gunicorn = "19.9.0"
Pillow = "5.2.0"
pylint = "2.1.1"
sentry-sdk = "0.4.3"
python-xmp-toolkit = "^2.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
